Hiding private data, make a user-task relationship and authenticating endpoints

+++ src/routers/task.js
@@ -1,23 +1,26 @@
 const express = require('express')
 const Task = require('../db/models/task')
+const auth = require('../middleware/auth')
 const router = new express.Router()
 
+router.post('/tasks', auth, async (req, res) => {
+    const task = new Task({
+        ...req.body,
+        owner: req.user._id
+    })
 
-router.post('/tasks', async (req, res) => {
-    const t = new Task(req.body)
     try {
-        await t.save()
-        res.status(201).send(t)
-    } catch(e) {
+        await task.save()
+        res.status(201).send(task)
+    } catch (e) {
         res.status(400).send(e)
     }
 })
 
 
-
-router.get('/tasks',async (req, res) => {
+router.get('/tasks', auth, async (req, res) => {
     try {
-        const tasks = await Task.find({})
+        const tasks = await Task.find({owner:req.user._id})
         res.send(tasks)
     }
     catch (e) {
@@ -25,10 +28,10 @@ router.get('/tasks',async (req, res) => {
     }   
 })
 
-router.get('/tasks/:id',async (req, res) => {
+router.get('/tasks/:id',auth,async (req, res) => {
     const _id = req.params.id
     try{
-        const task = await Task.findById({_id})
+        const task = await Task.findOne({_id, owner:req.user._id})
         if(!task) res.status(404).send
         res.send(task)
     } catch (e) {
@@ -37,14 +40,14 @@ router.get('/tasks/:id',async (req, res) => {
 })
 
 
-router.patch('/tasks/:id' , async (req, res) => {
+router.patch('/tasks/:id' ,auth,  async (req, res) => {
     const updates = Object.keys(req.body)
     const allowed = ["description", "completed"]
     const isValidOp = updates.every((update) => allowed.includes(update))
     if(!isValidOp) return res.status(404).send("Invalid updates")
     try {
         // const task = await Task.findByIdAndUpdate(req.params.id,req.body , {new: true, runValidators: true})
-        const task = await Task.findById(req.params.id)
+        const task = await Task.findOne({_id:req.params.id, owner:req.user.id})
         updates.forEach((update) => task[update] = req.body[update])
         await task.save()
         if(!task) return res.status(404).send()
@@ -54,9 +57,9 @@ router.patch('/tasks/:id' , async (req, res) => {
     }
 })
 
-router.delete('/tasks/:id' , async (req,res) => {
+router.delete('/tasks/:id' ,auth , async (req,res) => {
     try{
-        const task = await Task.findByIdAndDelete(req.params.id)
+        const task = await Task.findByIdAndDelete({_id:req.params.id, owner:req.user.id})
         if(!task) return res.status(404).send()
         res.send(task)
     } catch (e) {
